平台外壳搬出 data，data 层不再 import android.*
搬到新的 platform/ 包：KeystoreSecretCipher、DefaultImagePreparer、ApkInstaller +
ApkInstallResultReceiver（AndroidManifest 里的 receiver 名同步改）、以及打开数据库文件的
createNodeSeekDatabase。接口留在 data：SecretCipher、ImagePreparer、NodeSeekDatabase 本身。

AppUpdateRepository.onInstallStatus(Int) 改成 onInstallOutcome(InstallOutcome)。
哪个整数表示「用户点了取消」是 PackageInstaller 的事实，翻译放在广播到达的地方，
并补了这段映射的测试——之前只有 CONFLICT 一条路径被间接覆盖到。
